# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BackBet API",
    "version": "1.0.0",
    "description": "API de Apostas - Backend",
    "contact": {
      "name": "BackBet Team",
      "url": "https://github.com/dioneyfroes-coder/backBet"
    },
    "license": {
      "name": "ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Desenvolvimento"
    },
    {
      "url": "https://api.backbet.com/api",
      "description": "Produção"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token ou User ID em desenvolvimento"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "BAD_REQUEST"
              },
              "message": {
                "type": "string",
                "example": "Dados inválidos"
              },
              "details": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/ValidationError"
                  },
                  {
                    "type": "object"
                  }
                ]
              }
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "timestamp": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        },
        "example": {
          "success": false,
          "error": {
            "code": "BAD_REQUEST",
            "message": "Dados inválidos",
            "details": {
              "username": "Invalid input: expected string"
            }
          },
          "meta": {
            "timestamp": "2025-12-04T15:40:17.562Z"
          }
        }
      },
...
      "ValidationError": {
        "type": "object",
        "description": "Detalhes de validação por campo",
        "additionalProperties": {
          "type": "string",
          "example": "Campo obrigatório"
        },
        "example": {
          "email": "Formato de email inválido",
          "username": "Tamanho mínimo: 3"
        }
      },
...
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "example": "DEPOSIT"
          },
          "amount": {
            "type": "number",
            "example": 100.5
          },
          "currency": {
            "type": "string",
            "example": "BRL"
          },
          "description": {
            "type": "string",
            "example": "Depósito via cartão"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TransactionHistory": {
        "type": "object",
        "properties": {
          "transactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Transaction"
            }
          },
          "total": {
            "type": "integer",
            "example": 42
          }
        }
      },
      "UnauthorizedError": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "UNAUTHORIZED"
              },
              "message": {
                "type": "string",
                "example": "Autenticação requerida"
              }
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "timestamp": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        },
        "example": {
          "success": false,
          "error": {
            "code": "UNAUTHORIZED",
            "message": "Autenticação requerida"
          },
          "meta": {
            "timestamp": "2025-11-14T23:21:00.000Z"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/wallets/me": {
      "get": {
        "tags": [
          "Wallets"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Retorna a carteira do usuário autenticado",
        "responses": {
          "200": {
            "description": "Carteira encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletResponse"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "404": {
            "description": "Carteira não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/wallets/deposit": {
      "post": {
        "tags": [
          "Wallets"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Realiza depósito na carteira do usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number",
                    "minimum": 1,
                    "example": 100
                  },
                  "currency": {
                    "type": "string",
                    "enum": [
                      "BRL",
                      "USD",
                      "EUR"
                    ],
                    "example": "BRL"
                  },
                  "description": {
                    "type": "string",
                    "example": "Depósito via cartão de crédito"
                  }
                },
                "required": [
                  "amount",
                  "currency"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Depósito realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          }
        }
      }
    },
    "/api/wallets/withdraw": {
      "post": {
        "tags": [
          "Wallets"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Realiza saque da carteira do usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number",
                    "minimum": 100,
                    "example": 150
                  },
                  "currency": {
                    "type": "string",
                    "enum": [
                      "BRL",
                      "USD",
                      "EUR"
                    ],
                    "example": "BRL"
                  },
                  "description": {
                    "type": "string",
                    "example": "Saque para conta bancária"
                  }
                },
                "required": [
                  "amount",
                  "currency"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Saque realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos ou saldo insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          }
        }
      }
    },
    "/api/wallets/history": {
      "get": {
        "tags": [
          "Wallets"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Retorna histórico de transações da carteira",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Número de transações a retornar"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Deslocamento para paginação"
          }
        ],
        "responses": {
          "200": {
            "description": "Histórico retornado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionHistory"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          }
        }
      }
    }
  }
}